requirements.txt
pandas==1.1.5
six==1.12.0
tensorflow==2.2.0
stripe==2.27.0
deepface==0.0.49
opencv_contrib_python==4.5.2.54
matplotlib==3.3.3
Django>=4.0
numpy==1.16.0
wget==3.2
python_docx==0.8.10
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
coolname==1.1.0
object_detection==0.0.3
Pillow==8.3.2
psutil
